feat: split `useAnimaCodegen` hook
