+++ sdk-react/package.json
@@ -29,9 +29,7 @@
   ],
   "dependencies": {
     "eventsource": "^3.0.6",
-    "immer": "^10.1.1",
-    "swr": "^2.3.3",
-    "use-immer": "^0.11.0"
+    "swr": "^2.3.3"
   },
   "peerDependencies": {
     "@animaapp/anima-sdk": "workspace:*",
